improve getting outline accessor method, add error handling
